Two changes lumped together that because I forgot to commit.
The first: Add an UNLOCK_HACKS system so that tech prerequisites can be
hacked on to items that need them. The game deals with this a different
way, but it depends on the order that the items are laid out in the file
(not something stable like the ID ordering), so I'd rather just specify
it explicitly.

The second: Add a way to just list all the item and tech names.